[Refactor] OCP 위반을 전략패턴을 통해 해결 #58
