+++ src/main/java/server/inuappcenter/kr/controller/boardController/FaqController.java
@@ -2,15 +2,15 @@ package server.inuappcenter.kr.controller.boardController;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import server.inuappcenter.kr.common.data.dto.CommonResponseDto;
 import server.inuappcenter.kr.data.dto.request.FaqBoardRequestDto;
 import server.inuappcenter.kr.data.dto.response.BoardResponseDto;
-import server.inuappcenter.kr.service.boardService.AdditionalBoardService;
+import server.inuappcenter.kr.service.boardService.AdditionalBoardStrategyProvider;
 import server.inuappcenter.kr.service.boardService.BoardService;
 
 import javax.validation.Valid;
@@ -19,15 +19,11 @@ import java.util.List;
 @RestController
 @RequestMapping("/faqs")
 @Tag(name = "[FAQ] 질의응답 게시판")
+@RequiredArgsConstructor
 @Slf4j
 public class FaqController {
     private final BoardService boardService;
-    private final AdditionalBoardService additionalBoardService;
-
-    public FaqController(BoardService boardService, @Qualifier(value = "FaqBoardServiceImpl") AdditionalBoardService additionalBoardService) {
-        this.boardService = boardService;
-        this.additionalBoardService = additionalBoardService;
-    }
+    private final AdditionalBoardStrategyProvider additionalBoardStrategyProvider;
 
     @Operation(summary = "FAQ 한 개 가져오기", description = "조회할 faq_id를 입력해주세요")
     @GetMapping("public/{id}")
@@ -40,7 +36,7 @@ public class FaqController {
     public ResponseEntity<List<BoardResponseDto>> getFaqBoardList(
             @RequestParam(value = "topic", required = false) String topic
     ) {
-        return ResponseEntity.status(HttpStatus.OK).body(additionalBoardService.findBoardList(topic));
+        return ResponseEntity.status(HttpStatus.OK).body(additionalBoardStrategyProvider.findBoardList("FaqBoardServiceImpl", topic));
     }
 
     @Operation(summary = "FAQ 한 개 작성", description = "저장할 FAQ JSON을 보내주세요")
