+++ src/main/java/server/inuappcenter/kr/service/boardService/impl/FaqBoardServiceImpl.java
@@ -2,21 +2,19 @@ package server.inuappcenter.kr.service.boardService.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import server.inuappcenter.kr.data.domain.board.Board;
 import server.inuappcenter.kr.data.domain.board.FaqBoard;
 import server.inuappcenter.kr.data.dto.response.BoardResponseDto;
 import server.inuappcenter.kr.data.repository.FaqRepository;
-import server.inuappcenter.kr.service.boardService.AdditionalBoardService;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service("FaqBoardServiceImpl")
+@Service
 @RequiredArgsConstructor
 @Slf4j
 public class FaqBoardServiceImpl implements AdditionalBoardService {
