+++ src/main/java/server/inuappcenter/kr/service/boardService/impl/PhotoBoardServiceImpl.java
@@ -7,13 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import server.inuappcenter.kr.data.domain.board.Board;
 import server.inuappcenter.kr.data.dto.response.BoardResponseDto;
 import server.inuappcenter.kr.data.repository.PhotoBoardRepository;
-import server.inuappcenter.kr.service.boardService.AdditionalBoardService;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
-@Service("PhotoBoardServiceImpl")
+@Service
 @RequiredArgsConstructor
 @Slf4j
 public class PhotoBoardServiceImpl implements AdditionalBoardService {
